+++ Backend/src/main/java/transport/service/ChuyenXeService.java
@@ -1,6 +1,7 @@
 package transport.service;
 
 import org.springframework.stereotype.Service;
+import transport.dto.ChuyenXeRequest;
 import transport.model.ChuyenXe;
 
 import java.util.List;
@@ -9,6 +10,6 @@ import java.util.Optional;
 public interface ChuyenXeService {
     List<ChuyenXe> getAllChuyenXe();
     void saveChuyenXe(ChuyenXe chuyenXe);
-    void deleteChuyenXe(int id);
-    Optional<ChuyenXe> getChuyenXeById(int id);
+    void deleteChuyenXe(Long id);
+    Optional<ChuyenXe> getChuyenXeById(Long id);
 }
