+++ Backend/src/main/java/transport/service/implement/ChuyenXeServiceImpl.java
@@ -2,6 +2,7 @@ package transport.service.implement;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import transport.dto.ChuyenXeRequest;
 import transport.model.ChuyenXe;
 import transport.repository.ChuyenXeRepository;
 import transport.service.ChuyenXeService;
@@ -11,6 +12,7 @@ import java.util.Optional;
 
 @Service
 public class ChuyenXeServiceImpl implements ChuyenXeService {
+
     private ChuyenXeRepository chuyenXeRepo;
     @Autowired
     public ChuyenXeServiceImpl(ChuyenXeRepository chuyenXeRepo){ this.chuyenXeRepo = chuyenXeRepo; }
@@ -25,12 +27,12 @@ public class ChuyenXeServiceImpl implements ChuyenXeService {
     }
 
     @Override
-    public void deleteChuyenXe(int id) {
+    public void deleteChuyenXe(Long id) {
         chuyenXeRepo.deleteById(id);
     }
 
     @Override
-    public Optional<ChuyenXe> getChuyenXeById(int id) {
+    public Optional<ChuyenXe> getChuyenXeById(Long id) {
         return chuyenXeRepo.findById(id);
     }
 }
